Connect the consolidated studio data directly
To the modal to prevent constantly updating. Also only render the modals if they are visible
